pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lekha"
version = "0.1.0"
description = "CLI and web viewer for manuscript OCR comparison and correction."
authors = [{name = "Lekha Project"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "Flask>=3.0",
    "typer>=0.9",
    "pytesseract>=0.3",
    "Pillow>=10.0",
    "pdf2image>=1.16",
]

[project.urls]
Homepage = "https://example.com/lekha"

[project.scripts]
lekha = "lekha.cli:app"
